+++ hydra_physics/include/hydra/pathing/pathfinding.hpp
@@ -75,7 +75,7 @@ public:
 		//Must always be used to calculate G distance
 		float gDistanceTo(Node* nodeEnd)
 		{
-			return std::sqrt(std::pow(this->pos.x() - nodeEnd->pos.x(), 2.0f) + std::pow(this->pos.z() - nodeEnd->pos.z(), 2.0f));
+			return std::sqrt((float)std::pow(this->pos.x() - nodeEnd->pos.x(), 2.0f) + (float)std::pow(this->pos.z() - nodeEnd->pos.z(), 2.0f));
 		}
 		bool operator<(Node& other) { return this->getF() < other.getF(); }
 		bool operator==(Node& other) { return this->getF() == other.getF(); }
